# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AwayFromHighFilter": {
        "additionalProperties": false,
        "description": "\"Within X% of high\". ``100`` means ignore (docs/01 §2.4).",
        "properties": {
          "ath": {
            "default": 100,
            "title": "Ath",
            "type": "integer"
          },
          "one_year": {
            "default": 100,
            "title": "One Year",
            "type": "integer"
          }
        },
        "title": "AwayFromHighFilter",
        "type": "object"
      },
...
      "CircuitsFilter": {
        "additionalProperties": false,
        "description": "Maximum permitted circuit-hit days in the window. ``> 250`` means ignore (docs/01 §2.6).\n\nNote the asymmetry with the other sentinels: the ignore value is a *threshold*, not a single\nmagic number, because any cap above a full year of trading days cannot bind.",
        "properties": {
          "m1": {
            "default": 999,
            "title": "M1",
            "type": "integer"
          },
          "m12": {
            "default": 999,
            "title": "M12",
            "type": "integer"
          },
          "m3": {
            "default": 999,
            "title": "M3",
            "type": "integer"
          },
          "m6": {
            "default": 999,
            "title": "M6",
            "type": "integer"
          },
          "m9": {
            "default": 999,
            "title": "M9",
            "type": "integer"
          }
        },
        "title": "CircuitsFilter",
        "type": "object"
      },
...
      "CustomFilter": {
        "additionalProperties": false,
        "description": "A field-to-field comparison, e.g. ``ma_50 >= ma_200`` (docs/01 §2.14).",
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "left": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Left",
            "type": "string"
          },
          "op": {
            "enum": [
              ">=",
              "<=",
              "="
            ],
            "title": "Op",
            "type": "string"
          },
          "right": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Right",
            "type": "string"
          }
        },
        "required": [
          "left",
          "op",
          "right"
        ],
        "title": "CustomFilter",
        "type": "object"
      },
...
      "ExtraFactor": {
        "additionalProperties": false,
        "description": "Factor two / factor three of the combined ranking (docs/01 §2.12).",
        "properties": {
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "sort_by": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[a-z][a-z0-9_]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort By"
          },
          "sort_direction": {
            "default": "desc",
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Sort Direction",
            "type": "string"
          }
        },
        "title": "ExtraFactor",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "MovingAverageFilter": {
        "additionalProperties": false,
        "description": "Eight independent switches, AND-combined (docs/01 §2.3).",
        "properties": {
          "above_100": {
            "default": false,
            "title": "Above 100",
            "type": "boolean"
          },
          "above_20": {
            "default": false,
            "title": "Above 20",
            "type": "boolean"
          },
          "above_200": {
            "default": false,
            "title": "Above 200",
            "type": "boolean"
          },
          "above_50": {
            "default": false,
            "title": "Above 50",
            "type": "boolean"
          },
          "below_100": {
            "default": false,
            "title": "Below 100",
            "type": "boolean"
          },
          "below_20": {
            "default": false,
            "title": "Below 20",
            "type": "boolean"
          },
          "below_200": {
            "default": false,
            "title": "Below 200",
            "type": "boolean"
          },
          "below_50": {
            "default": false,
            "title": "Below 50",
            "type": "boolean"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "title": "MovingAverageFilter",
        "type": "object"
      },
      "PeFilter-Input": {
        "additionalProperties": false,
        "description": "P/E band behind its own switch. Undefined P/E is excluded when on (docs/01 §2.8).",
        "properties": {
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "from": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From"
          },
          "to": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To"
          }
        },
        "title": "PeFilter",
        "type": "object"
      },
...
      "PositiveDaysFilter": {
        "additionalProperties": false,
        "description": "Minimum % of trading days that closed up. ``0`` means ignore (docs/01 §2.5).",
        "properties": {
          "m1": {
            "default": 0,
            "title": "M1",
            "type": "integer"
          },
          "m12": {
            "default": 0,
            "title": "M12",
            "type": "integer"
          },
          "m3": {
            "default": 0,
            "title": "M3",
            "type": "integer"
          },
          "m6": {
            "default": 0,
            "title": "M6",
            "type": "integer"
          },
          "m9": {
            "default": 0,
            "title": "M9",
            "type": "integer"
          }
        },
        "title": "PositiveDaysFilter",
        "type": "object"
      },
...
      "ProblemOut": {
        "additionalProperties": true,
        "description": "RFC 9457. ``extra=\"allow\"`` carries the per-type members docs/07 attaches.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemOut",
        "type": "object"
      },
      "RangeFilter-Input": {
        "additionalProperties": false,
        "description": "An inclusive ``[from, to]`` band; both ``None`` means ignore (docs/06 step 4).",
        "properties": {
          "from": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From"
          },
          "to": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To"
          }
        },
        "title": "RangeFilter",
        "type": "object"
      },
...
      "RunRequest": {
        "additionalProperties": false,
        "description": "docs/07: `{ \"as_of\": \"2026-08-19\" | null, \"override_definition\": {…} | null }`.",
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "data_version": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Version"
          },
          "override_definition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScreenDefinition-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "RunRequest",
        "type": "object"
      },
...
      "ScreenDefinition-Input": {
        "additionalProperties": false,
        "description": "The persisted screen configuration.\n\nTwo definitions that mean the same thing produce the same :meth:`canonical_json`, and\ntherefore the same :meth:`definition_hash` — which is what makes the Redis cache key\n(docs/06 §Caching) and ``screen_run.definition_hash`` trustworthy.",
        "properties": {
          "apply_filters_on": {
            "default": "all",
            "enum": [
              "all",
              "decile_1",
              "decile_2",
              "decile_3",
              "decile_4",
              "decile_5",
              "top_50",
              "top_100"
            ],
            "title": "Apply Filters On",
            "type": "string"
          },
          "away_from_high": {
            "$ref": "#/components/schemas/AwayFromHighFilter"
          },
          "circuits": {
            "$ref": "#/components/schemas/CircuitsFilter"
          },
          "custom_filters": {
            "items": {
              "$ref": "#/components/schemas/CustomFilter"
            },
            "title": "Custom Filters",
            "type": "array"
          },
          "factor_three": {
            "$ref": "#/components/schemas/ExtraFactor"
          },
          "factor_two": {
            "$ref": "#/components/schemas/ExtraFactor"
          },
          "historical_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Historical Date"
          },
          "ignore_above_beta": {
            "default": 100,
            "title": "Ignore Above Beta",
            "type": "integer"
          },
          "ignore_top_beta": {
            "$ref": "#/components/schemas/TopRiskFilter"
          },
          "ignore_top_volatility": {
            "$ref": "#/components/schemas/TopRiskFilter"
          },
          "index": {
            "enum": [
              "nifty-50",
              "nifty-next-50",
              "nifty-100",
              "nifty-200",
              "nifty-500",
              "nifty-total-market",
              "nifty-large-mid-250",
              "nifty-midcap-150",
              "nifty-smallcap-250",
              "nifty-microcap-250",
              "nifty-mid-small-400",
              "nifty-allcap",
              "nifty-fno",
              "etf"
            ],
            "title": "Index",
            "type": "string"
          },
          "marketcap": {
            "$ref": "#/components/schemas/RangeFilter-Input"
          },
          "median_volume_1y": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Volume 1Y"
          },
          "min_return_1y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Return 1Y"
          },
          "moving_average": {
            "$ref": "#/components/schemas/MovingAverageFilter"
          },
          "pe": {
            "$ref": "#/components/schemas/PeFilter-Input"
          },
          "positive_days": {
            "$ref": "#/components/schemas/PositiveDaysFilter"
          },
          "price": {
            "$ref": "#/components/schemas/RangeFilter-Input"
          },
          "series": {
            "items": {
              "type": "string"
            },
            "title": "Series",
            "type": "array"
          },
          "sort_by": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Sort By",
            "type": "string"
          },
          "sort_direction": {
            "default": "desc",
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Sort Direction",
            "type": "string"
          }
        },
        "required": [
          "index",
          "sort_by"
        ],
        "title": "ScreenDefinition",
        "type": "object"
      },
...
      "ScreenRunResponse": {
        "additionalProperties": false,
        "description": "docs/07 §\"Running a screen\", field for field.",
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "data_version": {
            "title": "Data Version",
            "type": "integer"
          },
          "result_count": {
            "title": "Result Count",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ScreenRunRowOut"
            },
            "title": "Rows",
            "type": "array"
          },
          "sorting_factor": {
            "$ref": "#/components/schemas/SortingFactorOut"
          }
        },
        "required": [
          "as_of",
          "data_version",
          "result_count",
          "sorting_factor",
          "columns",
          "rows"
        ],
        "title": "ScreenRunResponse",
        "type": "object"
      },
      "ScreenRunRowOut": {
        "additionalProperties": true,
        "description": "One result row: ``rank`` plus one member per entry of ``columns``.\n\n``extra=\"allow\"`` because the column set is the user's, not the schema's — which is what\ndocs/01 §4's column picker means. In OpenAPI this becomes ``additionalProperties``, so the\ngenerated client types the known keys and permits the rest.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "sorting_factor": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "rank",
          "symbol",
          "name"
        ],
        "title": "ScreenRunRowOut",
        "type": "object"
      },
...
      "SortingFactorOut": {
        "additionalProperties": false,
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label"
        ],
        "title": "SortingFactorOut",
        "type": "object"
      },
...
      "TopRiskFilter": {
        "additionalProperties": false,
        "description": "\"Ignore Top Beta / Top Volatility\" — a precomputed per-universe flag (docs/06 step 4).",
        "properties": {
          "count": {
            "default": 0,
            "title": "Count",
            "type": "integer"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "title": "TopRiskFilter",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "India-equities momentum screener. See docs/07-api-spec.md.",
    "title": "Decile API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/screens/{public_id}/run": {
      "post": {
        "description": "docs/07 §\"Running a screen\".",
        "operationId": "runSavedScreen",
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "title": "Public Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Invalid screen definition"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Authentication required"
          },
          "402": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Your plan does not include this feature"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data version is no longer current"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "No trading day available for that date"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data pipeline is degraded"
          }
        },
        "summary": "Run a saved screen",
        "tags": [
          "screens"
        ]
      }
    },
...
  }
}